requirements.txt
python==3.9.5
pandas==1.1.3 
nltk==3.6.2 
gensim==4.0.1 
scikit-learn==0.24.2
fuzzywuzzy==0.18.0
mysql-connector-python==8.0.25
streamlit==0.85.1
requests==2.24.0
beautifulsoup4==4.9.3
selenium==3.141.0